+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 + [x] Keylock: added new component
 + [x] Calendar: fixed bug issue #1732
 + [x] Window: fixed event `onMaxClick`, issue #1735
++ [x] DoubleSelectBox: added new component, request from Patreon patron
 
 ### 4.4.3
 #### M4Q